move automatic removal to server page and view client components, and fix up storybook
