+++ src/app/(proper_react)/redesign/(authenticated)/user/dashboard/fix/data-broker-profiles/view-data-brokers/ViewDataBrokers.stories.tsx
@@ -114,7 +114,13 @@ const ViewWrapper = (props: ViewWrapperProps) => {
   };
 
   return (
-    <Shell l10n={getEnL10nSync()} session={mockedSession} nonce="">
+    <Shell
+      l10n={getEnL10nSync()}
+      session={mockedSession}
+      nonce=""
+      monthlySubscriptionUrl=""
+      yearlySubscriptionUrl=""
+    >
       <FixView
         breaches={mockedBreachesEmpty}
         userScannedResults={scanData.results}
